fix issues in card component and move it to a common as it's a general layout component
